Update: cleaned up header and added its own router
